refactor: isolate context search orchestration
